bot_requirements.txt
python-telegram-bot==8.1.1
sqlalchemy==1.1.14
psycopg2==2.7.3.1
